【Python】网易云音乐歌曲批量下载,免VIP【支持歌单】
【Python】网易云音乐歌曲批量下载,免VIP【支持歌单】

+++ Python3 网易云音乐mp3下载/downMusic.py
@@ -19,24 +19,23 @@ class HttpSvc():
 		downNum    = 0
 
 		for x in values:
-
-			print('***** ' + x['name'] + '.mp3 ***** Downloading...')
-			url = 'http://music.163.com/song/media/outer/url?id=' + x['id'] + '.mp3'
-			try:
-				urllib.request.urlretrieve(url,'d:/music/' + x['name'] + '.mp3')
-				downNum = downNum + 1
-			except:
-				x = x - 1
-				print('Download wrong~')
-
+			if not os.path.exists("d:/music/" + x['name'] + '.mp3'):
+				print('***** ' + x['name'] + '.mp3 ***** Downloading...')
+				url = 'http://music.163.com/song/media/outer/url?id=' + x['id'] + '.mp3'
+				try:
+					urllib.request.urlretrieve(url,'d:/music/' + x['name'] + '.mp3')
+					downNum = downNum + 1
+				except:
+					x = x - 1
+					print('Download wrong~')
 		print('Download complete ' + str(downNum) + 'files !')
 		pass
 
-	def getMusicData(self,id):
+	def getMusicData(self,url):
 
 		user_agent = 'Mozilla/5.0 (Windows NT 6.3; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/52.0.2743.82 Safari/537.36'
 		headers    = {'User-Agent':user_agent}
-		webData    = requests.get('https://music.163.com/playlist?id=' + id,headers=headers).text
+		webData    = requests.get(url,headers=headers).text
 		soup       = BeautifulSoup(webData,'lxml')
 
 		find_list  = soup.find('ul',class_="f-hide").find_all('a')
@@ -50,5 +49,6 @@ class HttpSvc():
 if __name__ == '__main__':
 
 	newHttp = HttpSvc()
-	musicData = newHttp.getMusicData('2074950566') #获取歌单歌曲id
+	musicData = newHttp.getMusicData('https://music.163.com/playlist?id=632061035') #获取歌单歌曲id https://music.163.com/playlist?id=2074950566
+	print(musicData)
 	print(newHttp.get(musicData))
